cleaning up and documenting block_acquisition
