+++ node/src/components/block_synchronizer/block_acquisition.rs
@@ -4,112 +4,66 @@ use std::{
 };
 
 use datasize::DataSize;
-use derive_more::{Display, From};
-use tracing::{debug, error, info, trace, warn};
+use derive_more::Display;
+use tracing::{debug, error, info, trace};
 
 use casper_hashing::Digest;
-use casper_types::{EraId, PublicKey};
-
-use super::deploy_acquisition;
+use casper_types::PublicKey;
 
 use crate::{
     components::block_synchronizer::{
-        deploy_acquisition::{DeployAcquisition, DeployIdentifier},
-        need_next::NeedNext,
-        peer_list::PeerList,
-        signature_acquisition::SignatureAcquisition,
+        block_acquisition_action::BlockAcquisitionAction, deploy_acquisition::DeployAcquisition,
+        peer_list::PeerList, signature_acquisition::SignatureAcquisition, BlockAcquisitionError,
         ExecutionResultsAcquisition, ExecutionResultsChecksum,
     },
     types::{
-        ApprovalsHashes, Block, BlockExecutionResultsOrChunk, BlockExecutionResultsOrChunkId,
-        BlockHash, BlockHeader, DeployHash, DeployId, EraValidatorWeights, FinalitySignature, Item,
-        NodeId, SignatureWeight,
+        ApprovalsHashes, Block, BlockExecutionResultsOrChunk, BlockHash, BlockHeader, DeployHash,
+        DeployId, EraValidatorWeights, FinalitySignature, Item, SignatureWeight,
     },
     NodeRng,
 };
 
-#[allow(dead_code)] // todo! do a pass on error variants
-#[derive(Clone, Copy, From, PartialEq, Eq, DataSize, Debug)]
-pub(crate) enum Error {
-    InvalidStateTransition,
-    BlockHashMismatch {
-        expected: BlockHash,
-        actual: BlockHash,
-    },
-    RootHashMismatch {
-        expected: Digest,
-        actual: Digest,
-    },
-    #[from]
-    InvalidAttemptToApplyApprovalsHashes(deploy_acquisition::Error),
-    InvalidAttemptToApplyGlobalState {
-        root_hash: Digest,
-    },
-    InvalidAttemptToApplyDeploy {
-        deploy_id: DeployId,
-    },
-    InvalidAttemptToApplyExecutionResults,
-    InvalidAttemptToApplyExecutionResultsChecksum,
-    InvalidAttemptToApplyStoredExecutionResults,
-    InvalidAttemptToAcquireExecutionResults,
-    InvalidAttemptToMarkComplete,
-    ExecutionResults(super::execution_results_acquisition::Error),
-}
-
-impl Display for Error {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Error::InvalidStateTransition => write!(f, "invalid state transition"),
-            Error::InvalidAttemptToMarkComplete => {
-                write!(f, "invalid attempt to mark complete")
-            }
-            Error::InvalidAttemptToApplyGlobalState { root_hash } => {
-                write!(
-                    f,
-                    "invalid attempt to apply invalid global hash root hash: {}",
-                    root_hash
-                )
-            }
-            Error::InvalidAttemptToApplyDeploy { deploy_id } => {
-                write!(f, "invalid attempt to apply invalid deploy {}", deploy_id)
-            }
-            Error::InvalidAttemptToApplyExecutionResults => {
-                write!(f, "invalid attempt to apply execution results")
-            }
-            Error::InvalidAttemptToApplyExecutionResultsChecksum => {
-                write!(f, "invalid attempt to apply execution results checksum")
-            }
-            Error::InvalidAttemptToApplyStoredExecutionResults => {
-                write!(f, "invalid attempt to apply stored execution results notification; execution results are not in terminal state")
-            }
-            Error::InvalidAttemptToAcquireExecutionResults => {
-                write!(
-                    f,
-                    "invalid attempt to acquire execution results while in a terminal state"
-                )
-            }
-            Error::BlockHashMismatch { expected, actual } => {
-                write!(
-                    f,
-                    "block hash mismatch: expected {} actual: {}",
-                    expected, actual
-                )
-            }
-            Error::RootHashMismatch { expected, actual } => write!(
-                f,
-                "root hash mismatch: expected {} actual: {}",
-                expected, actual
-            ),
-            Error::ExecutionResults(error) => write!(f, "execution results error: {}", error),
-            Error::InvalidAttemptToApplyApprovalsHashes(error) => write!(
-                f,
-                "invalid attempt to apply approvals hashes results: {}",
-                error
-            ),
-        }
-    }
-}
-
+// BlockAcquisitionState is milestone oriented state machine; it is always in a resting state
+// indicating the last completed step, while attempting to acquire the necessary data to transition
+// to the next resting state milestone. the start and end of the workflow is fairly linear, but the
+// middle steps conditionally branch depending upon if this is a historical block (needs execution
+// state) or a block we intend to execute, and if the block body has one or more deploys.
+//
+// the state transitions are in a particular order, but there are divergent paths
+// based upon whether we are attempting to acquire global state and execution effects
+// (collectively referred to as execution state), and whether this particular block
+// contains any deploys
+//
+// blocks always require a header & body and sufficient finality signatures; blocks may contain
+// one or more deploys. if a block has any deploys, we must also acquire execution effects
+// for the deploys in the block (we do this as a chunked aggregate), and for post 1.5 blocks
+// we must also acquire approvals hashes (which correlate to which authorized account holders
+// signed the deploys).
+//
+// there are two levels of finality, weak and strict. we first get the bare minimum data about
+// a block, and then attempt to acquire at least weak finality before doing further work
+// acquiring data for a block, to avoid being tricked into wasting resources downloading bogus
+// blocks. with at least weak finality, we can relatively safely go about acquiring the rest
+// of the block's required records. if we have not acquired strong finality by the time we've
+// downloaded everything else, we do another round of asking for remaining signatures before
+// accepting the sync'd block.
+//
+// when acquiring data for a historical block, we want global state (always) and execution
+// effects (if any). when acquiring sufficient data to execute a block, we do not acquired
+// global state or execution effects. however, we still check for existence of an execution
+// effect _checksum_ in global state as an expedient way to determine if a block was
+// created post-1.5
+//
+// note that we check local storage prior to deciding to ask peers for records, as we may have
+// received records we need via gossiping. similarly, we collect finality signatures during each
+// state between HaveBlockHeader and HaveStrictFinalitySignatures inclusive and may acquire
+// strict finality before we check for it at the very end. finally due to the trie store nature
+// of global state, other than the first downloaded historical block we likely have the vast
+// majority of global state data locally already. for these reasons, it is common for most
+// blocks to transition thru the various states very quickly...particularly blocks with deploys.
+// however, the first block downloaded or blocks with a lot of deploys and / or execution state
+// churn can take arbitrarily longer on their relevant steps.
+//                                          ¯\_(ツ)_/¯
 #[derive(Clone, DataSize, Debug)]
 pub(super) enum BlockAcquisitionState {
     Initialized(BlockHash, SignatureAcquisition),
@@ -205,6 +159,178 @@ pub(super) enum Acceptance {
 }
 
 impl BlockAcquisitionState {
+    // the BlockAcquisitionState states and their valid transitions follow:
+    //
+    //   Initialized -> need block header
+    //
+    //   HaveBlockHeader -> if no era validators -> need era validator weights
+    //                    else need weak finality
+    //
+    //   HaveWeakFinalitySignatures -> need block
+    //
+    //   HaveBlock -> if should_fetch_execution_state -> need global state
+    //              else if block has deploys need approvals hashes
+    //              else if no deploys need strict finality
+    //
+    //   HaveGlobalState -> if should_fetch_execution_state
+    //                      if block has deploys ->
+    //                       if have execution effects -> need approvals hashes
+    //                       else -> need execution effects
+    //                      else -> need strict finality
+    //                     else -> error
+    //
+    //   HaveAllExecutionResults -> if should_fetch_execution_state
+    //                                if approvals checkable -> need approvals hashes
+    //                                else -> need deploys
+    //                               else error
+    //
+    //   HaveApprovalsHashes -> need deploys
+    //
+    //   HaveDeploys -> need strict finality
+    //
+    //   HaveStrictFinalitySignatures -> HaveStrictFinalitySignatures (success / terminal)
+    //
+    //   Failed -> Failed (terminal)
+    /// Determines what action should be taken to acquire the next needed block related data.
+    pub(super) fn next_action(
+        &mut self,
+        peer_list: &PeerList,
+        validator_weights: &EraValidatorWeights,
+        rng: &mut NodeRng,
+        should_fetch_execution_state: bool,
+    ) -> Result<BlockAcquisitionAction, BlockAcquisitionError> {
+        // self is the resting state we are in, ret is the next action that should be taken
+        // to acquire the necessary data to get us to the next step (if any), or an error
+        let ret = match self {
+            BlockAcquisitionState::Initialized(block_hash, ..) => Ok(
+                BlockAcquisitionAction::block_header(peer_list, rng, *block_hash),
+            ),
+            BlockAcquisitionState::HaveBlockHeader(block_header, signatures) => {
+                if validator_weights.is_empty() {
+                    Ok(BlockAcquisitionAction::era_validators(
+                        validator_weights.era_id(),
+                    ))
+                } else {
+                    Ok(BlockAcquisitionAction::finality_signatures(
+                        peer_list,
+                        rng,
+                        block_header,
+                        validator_weights
+                            .missing_validators(signatures.have_signatures())
+                            .cloned()
+                            .collect(),
+                    ))
+                }
+            }
+            BlockAcquisitionState::HaveWeakFinalitySignatures(header, _) => Ok(
+                BlockAcquisitionAction::block_body(peer_list, rng, header.id()),
+            ),
+            BlockAcquisitionState::HaveBlock(block, signatures, deploy_state) => {
+                if should_fetch_execution_state {
+                    Ok(BlockAcquisitionAction::global_state(
+                        peer_list,
+                        rng,
+                        *block.hash(),
+                        *block.state_root_hash(),
+                    ))
+                } else if deploy_state.needs_deploy().is_some() {
+                    Ok(BlockAcquisitionAction::approvals_hashes(
+                        block, peer_list, rng,
+                    ))
+                } else {
+                    Ok(BlockAcquisitionAction::strict_finality_signatures(
+                        peer_list,
+                        rng,
+                        block.header(),
+                        validator_weights,
+                        signatures,
+                    ))
+                }
+            }
+            BlockAcquisitionState::HaveGlobalState(
+                block,
+                signatures,
+                deploy_state,
+                exec_results,
+            ) => {
+                if should_fetch_execution_state == false {
+                    Err(BlockAcquisitionError::InvalidStateTransition)
+                } else if deploy_state.needs_deploy().is_some() {
+                    BlockAcquisitionAction::maybe_execution_results(
+                        block,
+                        peer_list,
+                        rng,
+                        exec_results,
+                    )
+                } else {
+                    Ok(BlockAcquisitionAction::strict_finality_signatures(
+                        peer_list,
+                        rng,
+                        block.header(),
+                        validator_weights,
+                        signatures,
+                    ))
+                }
+            }
+            BlockAcquisitionState::HaveAllExecutionResults(
+                block,
+                signatures,
+                deploys,
+                checksum,
+            ) => match (should_fetch_execution_state, checksum) {
+                (false, _) => Err(BlockAcquisitionError::InvalidStateTransition),
+                (true, ExecutionResultsChecksum::ApprovalsCheckable(_)) => Ok(
+                    BlockAcquisitionAction::approvals_hashes(block, peer_list, rng),
+                ),
+                (true, ExecutionResultsChecksum::ApprovalsUncheckable) => {
+                    Ok(BlockAcquisitionAction::maybe_needs_deploy(
+                        block.header(),
+                        peer_list,
+                        rng,
+                        validator_weights,
+                        signatures,
+                        deploys.needs_deploy(),
+                    ))
+                }
+            },
+            BlockAcquisitionState::HaveApprovalsHashes(block, signatures, deploys) => {
+                Ok(BlockAcquisitionAction::maybe_needs_deploy(
+                    block.header(),
+                    peer_list,
+                    rng,
+                    validator_weights,
+                    signatures,
+                    deploys.needs_deploy(),
+                ))
+            }
+            BlockAcquisitionState::HaveAllDeploys(header, signatures) => {
+                Ok(BlockAcquisitionAction::strict_finality_signatures(
+                    peer_list,
+                    rng,
+                    header.as_ref(),
+                    validator_weights,
+                    signatures,
+                ))
+            }
+            BlockAcquisitionState::HaveStrictFinalitySignatures(header, ..) => {
+                Ok(BlockAcquisitionAction::need_nothing(header.block_hash()))
+            }
+            BlockAcquisitionState::Failed(block_hash, ..) => {
+                Ok(BlockAcquisitionAction::need_nothing(*block_hash))
+            }
+        };
+        match &ret {
+            Ok(next_action) => {
+                info!(%next_action, %self);
+            }
+            Err(err) => {
+                error!(%err, %self);
+            }
+        }
+        ret
+    }
+
+    /// The block height of the current block, if available.
     pub(super) fn block_height(&self) -> Option<u64> {
         match self {
             BlockAcquisitionState::Initialized(..) | BlockAcquisitionState::Failed(..) => None,
@@ -221,14 +347,18 @@ impl BlockAcquisitionState {
         }
     }
 
-    pub(super) fn register_header(&mut self, header: BlockHeader) -> Result<Acceptance, Error> {
+    /// Register the block header for this block.
+    pub(super) fn register_block_header(
+        &mut self,
+        header: BlockHeader,
+    ) -> Result<Acceptance, BlockAcquisitionError> {
         let new_state = match self {
             BlockAcquisitionState::Initialized(block_hash, signatures) => {
                 if header.id() == *block_hash {
                     info!("BlockAcquisition: registering header for: {}", block_hash);
                     BlockAcquisitionState::HaveBlockHeader(Box::new(header), signatures.clone())
                 } else {
-                    return Err(Error::BlockHashMismatch {
+                    return Err(BlockAcquisitionError::BlockHashMismatch {
                         expected: *block_hash,
                         actual: header.id(),
                     });
@@ -250,71 +380,18 @@ impl BlockAcquisitionState {
         Ok(Acceptance::NeededIt)
     }
 
-    pub(super) fn register_approvals_hashes(
-        &mut self,
-        approvals_hashes: &ApprovalsHashes,
-        need_execution_state: bool,
-    ) -> Result<Acceptance, Error> {
-        let new_state = match self {
-            BlockAcquisitionState::HaveBlock(block, signatures, acquired)
-                if !need_execution_state =>
-            {
-                info!(
-                    "BlockAcquisition: registering approvals hashes for: {}",
-                    block.id()
-                );
-                acquired.apply_approvals_hashes(approvals_hashes)?;
-                BlockAcquisitionState::HaveApprovalsHashes(
-                    block.clone(),
-                    signatures.clone(),
-                    acquired.clone(),
-                )
-            }
-
-            BlockAcquisitionState::HaveAllExecutionResults(block, signatures, deploys, _)
-                if need_execution_state =>
-            {
-                deploys.apply_approvals_hashes(approvals_hashes)?;
-                info!(
-                    "BlockAcquisition: registering approvals hashes for: {}",
-                    block.id()
-                );
-                BlockAcquisitionState::HaveApprovalsHashes(
-                    block.clone(),
-                    signatures.clone(),
-                    deploys.clone(),
-                )
-            }
-            // we never ask for deploys in the following states, and thus it is erroneous to attempt
-            // to apply any
-            BlockAcquisitionState::HaveBlock(..)
-            | BlockAcquisitionState::HaveGlobalState(..)
-            | BlockAcquisitionState::Initialized(..)
-            | BlockAcquisitionState::HaveBlockHeader(..)
-            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
-            | BlockAcquisitionState::HaveAllExecutionResults(..)
-            | BlockAcquisitionState::HaveAllDeploys(..)
-            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
-            | BlockAcquisitionState::HaveApprovalsHashes(_, _, _)
-            | BlockAcquisitionState::Failed(..) => {
-                return Ok(Acceptance::HadIt);
-            }
-        };
-        self.set_state(new_state);
-        Ok(Acceptance::NeededIt)
-    }
-
+    /// Register the block body for this block.
     pub(super) fn register_block(
         &mut self,
         block: &Block,
         need_execution_state: bool,
-    ) -> Result<Acceptance, Error> {
+    ) -> Result<Acceptance, BlockAcquisitionError> {
         let new_state = match self {
             BlockAcquisitionState::HaveWeakFinalitySignatures(header, signatures) => {
                 let expected_block_hash = header.block_hash();
                 let actual_block_hash = block.hash();
                 if *actual_block_hash != expected_block_hash {
-                    return Err(Error::BlockHashMismatch {
+                    return Err(BlockAcquisitionError::BlockHashMismatch {
                         expected: expected_block_hash,
                         actual: *actual_block_hash,
                     });
@@ -349,11 +426,16 @@ impl BlockAcquisitionState {
         Ok(Acceptance::NeededIt)
     }
 
+    /// Register a finality signature for this block.
     pub(super) fn register_finality_signature(
         &mut self,
         signature: FinalitySignature,
         validator_weights: &EraValidatorWeights,
-    ) -> Result<Acceptance, Error> {
+    ) -> Result<Acceptance, BlockAcquisitionError> {
+        // we will accept finality signatures we don't yet have while in every state other than
+        // Initialized and Failed. However, it can only cause a state transition when we
+        // are in a resting state that needs weak finality or strict finality.
+
         let signer = signature.public_key.clone();
         let mut added = false;
         let mut maybe_block_hash: Option<BlockHash> = None;
@@ -408,11 +490,255 @@ impl BlockAcquisitionState {
         Ok(ret)
     }
 
+    /// Register the approvals hashes for this block.
+    pub(super) fn register_approvals_hashes(
+        &mut self,
+        approvals_hashes: &ApprovalsHashes,
+        need_execution_state: bool,
+    ) -> Result<Acceptance, BlockAcquisitionError> {
+        let new_state = match self {
+            BlockAcquisitionState::HaveBlock(block, signatures, acquired)
+                if !need_execution_state =>
+            {
+                info!(
+                    "BlockAcquisition: registering approvals hashes for: {}",
+                    block.id()
+                );
+                acquired.apply_approvals_hashes(approvals_hashes)?;
+                BlockAcquisitionState::HaveApprovalsHashes(
+                    block.clone(),
+                    signatures.clone(),
+                    acquired.clone(),
+                )
+            }
+
+            BlockAcquisitionState::HaveAllExecutionResults(block, signatures, deploys, _)
+                if need_execution_state =>
+            {
+                deploys.apply_approvals_hashes(approvals_hashes)?;
+                info!(
+                    "BlockAcquisition: registering approvals hashes for: {}",
+                    block.id()
+                );
+                BlockAcquisitionState::HaveApprovalsHashes(
+                    block.clone(),
+                    signatures.clone(),
+                    deploys.clone(),
+                )
+            }
+            // we never ask for deploys in the following states, and thus it is erroneous to attempt
+            // to apply any
+            BlockAcquisitionState::HaveBlock(..)
+            | BlockAcquisitionState::HaveGlobalState(..)
+            | BlockAcquisitionState::Initialized(..)
+            | BlockAcquisitionState::HaveBlockHeader(..)
+            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
+            | BlockAcquisitionState::HaveAllExecutionResults(..)
+            | BlockAcquisitionState::HaveAllDeploys(..)
+            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
+            | BlockAcquisitionState::HaveApprovalsHashes(_, _, _)
+            | BlockAcquisitionState::Failed(..) => {
+                return Ok(Acceptance::HadIt);
+            }
+        };
+        self.set_state(new_state);
+        Ok(Acceptance::NeededIt)
+    }
+
+    /// Register global state for this block.
+    pub(super) fn register_global_state(
+        &mut self,
+        root_hash: Digest,
+        need_execution_state: bool,
+    ) -> Result<(), BlockAcquisitionError> {
+        let new_state = match self {
+            BlockAcquisitionState::HaveBlock(block, signatures, deploys)
+                if need_execution_state =>
+            {
+                info!(
+                    "BlockAcquisition: registering global state for: {}",
+                    block.id()
+                );
+                if block.state_root_hash() == &root_hash {
+                    let block_hash = *block.hash();
+                    BlockAcquisitionState::HaveGlobalState(
+                        block.clone(),
+                        signatures.clone(),
+                        deploys.clone(),
+                        ExecutionResultsAcquisition::Needed { block_hash },
+                    )
+                } else {
+                    return Err(BlockAcquisitionError::RootHashMismatch {
+                        expected: *block.state_root_hash(),
+                        actual: root_hash,
+                    });
+                }
+            }
+            // we never ask for global state in the following states, and thus it is erroneous to
+            // attempt to apply any
+            BlockAcquisitionState::HaveBlock(..)
+            | BlockAcquisitionState::Initialized(..)
+            | BlockAcquisitionState::HaveBlockHeader(..)
+            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
+            | BlockAcquisitionState::HaveGlobalState(..)
+            | BlockAcquisitionState::HaveAllExecutionResults(..)
+            | BlockAcquisitionState::HaveAllDeploys(..)
+            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
+            | BlockAcquisitionState::HaveApprovalsHashes(..)
+            | BlockAcquisitionState::Failed(..) => {
+                return Ok(());
+            }
+        };
+        self.set_state(new_state);
+        Ok(())
+    }
+
+    /// Register execution results checksum for this block.
+    pub(super) fn register_execution_results_checksum(
+        &mut self,
+        execution_results_checksum: ExecutionResultsChecksum,
+        need_execution_state: bool,
+    ) -> Result<(), BlockAcquisitionError> {
+        match self {
+            BlockAcquisitionState::HaveGlobalState(
+                block,
+                _,
+                _,
+                acq @ ExecutionResultsAcquisition::Needed { .. },
+            ) if need_execution_state => {
+                info!(
+                    "BlockAcquisition: registering execution results hash for: {}",
+                    block.id()
+                );
+                *acq = acq
+                    .clone()
+                    .apply_checksum(execution_results_checksum)
+                    .map_err(BlockAcquisitionError::ExecutionResults)?;
+            }
+            BlockAcquisitionState::HaveAllDeploys(..)
+            | BlockAcquisitionState::HaveGlobalState(..)
+            | BlockAcquisitionState::HaveBlock(..)
+            | BlockAcquisitionState::Initialized(..)
+            | BlockAcquisitionState::HaveBlockHeader(..)
+            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
+            | BlockAcquisitionState::HaveAllExecutionResults(..)
+            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
+            | BlockAcquisitionState::HaveApprovalsHashes(..)
+            | BlockAcquisitionState::Failed(..) => {
+                return Ok(());
+            }
+        };
+        Ok(())
+    }
+
+    /// Register execution results or chunk for this block.
+    pub(super) fn register_execution_results_or_chunk(
+        &mut self,
+        block_execution_results_or_chunk: BlockExecutionResultsOrChunk,
+        need_execution_state: bool,
+    ) -> Result<Option<HashMap<DeployHash, casper_types::ExecutionResult>>, BlockAcquisitionError>
+    {
+        let (new_state, ret) = match self {
+            BlockAcquisitionState::HaveGlobalState(
+                block,
+                signatures,
+                deploys,
+                exec_results_acq,
+            ) if need_execution_state => {
+                info!(
+                    "BlockAcquisition: registering execution result or chunk for: {}",
+                    block.id()
+                );
+                let deploy_hashes = block.deploy_and_transfer_hashes().copied().collect();
+                match exec_results_acq
+                    .clone()
+                    .apply_block_execution_results_or_chunk(
+                        block_execution_results_or_chunk,
+                        deploy_hashes,
+                    ) {
+                    Ok(new_effects) => match new_effects {
+                        ExecutionResultsAcquisition::Needed { .. }
+                        | ExecutionResultsAcquisition::Pending { .. }
+                        | ExecutionResultsAcquisition::Acquiring { .. } => return Ok(None),
+                        ExecutionResultsAcquisition::Complete { ref results, .. } => (
+                            BlockAcquisitionState::HaveGlobalState(
+                                block.clone(),
+                                signatures.clone(),
+                                deploys.clone(),
+                                new_effects.clone(),
+                            ),
+                            Some(results.clone()),
+                        ),
+                    },
+                    Err(error) => {
+                        error!(%error, "failed to apply execution results");
+                        return Ok(None);
+                    }
+                }
+            }
+            BlockAcquisitionState::HaveAllDeploys(..)
+            | BlockAcquisitionState::HaveGlobalState(..)
+            | BlockAcquisitionState::HaveBlock(..)
+            | BlockAcquisitionState::Initialized(..)
+            | BlockAcquisitionState::HaveBlockHeader(..)
+            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
+            | BlockAcquisitionState::HaveAllExecutionResults(..)
+            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
+            | BlockAcquisitionState::HaveApprovalsHashes(..)
+            | BlockAcquisitionState::Failed(..) => {
+                return Ok(None);
+            }
+        };
+        self.set_state(new_state);
+        Ok(ret)
+    }
+
+    /// Register execution results stored notification for this block.
+    pub(super) fn register_execution_results_stored_notification(
+        &mut self,
+        need_execution_state: bool,
+    ) -> Result<(), BlockAcquisitionError> {
+        let new_state = match self {
+            BlockAcquisitionState::HaveGlobalState(
+                block,
+                signatures,
+                deploys,
+                ExecutionResultsAcquisition::Complete { checksum, .. },
+            ) if need_execution_state => {
+                info!(
+                    "BlockAcquisition: registering execution results stored notification for: {}",
+                    block.id()
+                );
+                BlockAcquisitionState::HaveAllExecutionResults(
+                    block.clone(),
+                    signatures.clone(),
+                    deploys.clone(),
+                    *checksum,
+                )
+            }
+            BlockAcquisitionState::HaveGlobalState(..)
+            | BlockAcquisitionState::HaveAllDeploys(..)
+            | BlockAcquisitionState::HaveBlock(..)
+            | BlockAcquisitionState::Initialized(..)
+            | BlockAcquisitionState::HaveBlockHeader(..)
+            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
+            | BlockAcquisitionState::HaveAllExecutionResults(..)
+            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
+            | BlockAcquisitionState::HaveApprovalsHashes(..)
+            | BlockAcquisitionState::Failed(..) => {
+                return Ok(());
+            }
+        };
+        self.set_state(new_state);
+        Ok(())
+    }
+
+    /// Register a deploy for this block.
     pub(super) fn register_deploy(
         &mut self,
         deploy_id: DeployId,
         need_execution_state: bool,
-    ) -> Result<Acceptance, Error> {
+    ) -> Result<Acceptance, BlockAcquisitionError> {
         let (new_state, acceptance) = match self {
             BlockAcquisitionState::HaveBlock(block, signatures, deploys)
                 if !need_execution_state =>
@@ -469,193 +795,11 @@ impl BlockAcquisitionState {
         Ok(acceptance)
     }
 
-    pub(super) fn register_global_state(
-        &mut self,
-        root_hash: Digest,
-        need_execution_state: bool,
-    ) -> Result<(), Error> {
-        let new_state = match self {
-            BlockAcquisitionState::HaveBlock(block, signatures, deploys)
-                if need_execution_state =>
-            {
-                info!(
-                    "BlockAcquisition: registering global state for: {}",
-                    block.id()
-                );
-                if block.state_root_hash() == &root_hash {
-                    let block_hash = *block.hash();
-                    BlockAcquisitionState::HaveGlobalState(
-                        block.clone(),
-                        signatures.clone(),
-                        deploys.clone(),
-                        ExecutionResultsAcquisition::Needed { block_hash },
-                    )
-                } else {
-                    return Err(Error::RootHashMismatch {
-                        expected: *block.state_root_hash(),
-                        actual: root_hash,
-                    });
-                }
-            }
-            // we never ask for global state in the following states, and thus it is erroneous to
-            // attempt to apply any
-            BlockAcquisitionState::HaveBlock(..)
-            | BlockAcquisitionState::Initialized(..)
-            | BlockAcquisitionState::HaveBlockHeader(..)
-            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
-            | BlockAcquisitionState::HaveGlobalState(..)
-            | BlockAcquisitionState::HaveAllExecutionResults(..)
-            | BlockAcquisitionState::HaveAllDeploys(..)
-            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
-            | BlockAcquisitionState::HaveApprovalsHashes(..)
-            | BlockAcquisitionState::Failed(..) => {
-                return Ok(());
-            }
-        };
-        self.set_state(new_state);
-        Ok(())
-    }
-
-    pub(super) fn register_execution_results_checksum(
-        &mut self,
-        execution_results_checksum: ExecutionResultsChecksum,
-        need_execution_state: bool,
-    ) -> Result<(), Error> {
-        match self {
-            BlockAcquisitionState::HaveGlobalState(
-                block,
-                _,
-                _,
-                acq @ ExecutionResultsAcquisition::Needed { .. },
-            ) if need_execution_state => {
-                info!(
-                    "BlockAcquisition: registering execution results hash for: {}",
-                    block.id()
-                );
-                *acq = acq
-                    .clone()
-                    .apply_checksum(execution_results_checksum)
-                    .map_err(Error::ExecutionResults)?;
-            }
-            BlockAcquisitionState::HaveAllDeploys(..)
-            | BlockAcquisitionState::HaveGlobalState(..)
-            | BlockAcquisitionState::HaveBlock(..)
-            | BlockAcquisitionState::Initialized(..)
-            | BlockAcquisitionState::HaveBlockHeader(..)
-            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
-            | BlockAcquisitionState::HaveAllExecutionResults(..)
-            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
-            | BlockAcquisitionState::HaveApprovalsHashes(..)
-            | BlockAcquisitionState::Failed(..) => {
-                return Ok(());
-            }
-        };
-        Ok(())
-    }
-
-    pub(super) fn register_execution_results_or_chunk(
-        &mut self,
-        block_execution_results_or_chunk: BlockExecutionResultsOrChunk,
-        need_execution_state: bool,
-    ) -> Result<Option<HashMap<DeployHash, casper_types::ExecutionResult>>, Error> {
-        let (new_state, ret) = match self {
-            BlockAcquisitionState::HaveGlobalState(
-                block,
-                signatures,
-                deploys,
-                exec_results_acq,
-            ) if need_execution_state => {
-                info!(
-                    "BlockAcquisition: registering execution result or chunk for: {}",
-                    block.id()
-                );
-                let deploy_hashes = block.deploy_and_transfer_hashes().copied().collect();
-                match exec_results_acq
-                    .clone()
-                    .apply_block_execution_results_or_chunk(
-                        block_execution_results_or_chunk,
-                        deploy_hashes,
-                    ) {
-                    Ok(new_effects) => match new_effects {
-                        ExecutionResultsAcquisition::Needed { .. }
-                        | ExecutionResultsAcquisition::Pending { .. }
-                        | ExecutionResultsAcquisition::Acquiring { .. } => return Ok(None),
-                        ExecutionResultsAcquisition::Complete { ref results, .. } => (
-                            BlockAcquisitionState::HaveGlobalState(
-                                block.clone(),
-                                signatures.clone(),
-                                deploys.clone(),
-                                new_effects.clone(),
-                            ),
-                            Some(results.clone()),
-                        ),
-                    },
-                    Err(error) => {
-                        error!(%error, "failed to apply execution results");
-                        return Ok(None);
-                    }
-                }
-            }
-            BlockAcquisitionState::HaveAllDeploys(..)
-            | BlockAcquisitionState::HaveGlobalState(..)
-            | BlockAcquisitionState::HaveBlock(..)
-            | BlockAcquisitionState::Initialized(..)
-            | BlockAcquisitionState::HaveBlockHeader(..)
-            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
-            | BlockAcquisitionState::HaveAllExecutionResults(..)
-            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
-            | BlockAcquisitionState::HaveApprovalsHashes(..)
-            | BlockAcquisitionState::Failed(..) => {
-                return Ok(None);
-            }
-        };
-        self.set_state(new_state);
-        Ok(ret)
-    }
-
-    pub(super) fn register_execution_results_stored_notification(
-        &mut self,
-        need_execution_state: bool,
-    ) -> Result<(), Error> {
-        let new_state = match self {
-            BlockAcquisitionState::HaveGlobalState(
-                block,
-                signatures,
-                deploys,
-                ExecutionResultsAcquisition::Complete { checksum, .. },
-            ) if need_execution_state => {
-                info!(
-                    "BlockAcquisition: registering execution results stored notification for: {}",
-                    block.id()
-                );
-                BlockAcquisitionState::HaveAllExecutionResults(
-                    block.clone(),
-                    signatures.clone(),
-                    deploys.clone(),
-                    *checksum,
-                )
-            }
-            BlockAcquisitionState::HaveGlobalState(..)
-            | BlockAcquisitionState::HaveAllDeploys(..)
-            | BlockAcquisitionState::HaveBlock(..)
-            | BlockAcquisitionState::Initialized(..)
-            | BlockAcquisitionState::HaveBlockHeader(..)
-            | BlockAcquisitionState::HaveWeakFinalitySignatures(..)
-            | BlockAcquisitionState::HaveAllExecutionResults(..)
-            | BlockAcquisitionState::HaveStrictFinalitySignatures(..)
-            | BlockAcquisitionState::HaveApprovalsHashes(..)
-            | BlockAcquisitionState::Failed(..) => {
-                return Ok(());
-            }
-        };
-        self.set_state(new_state);
-        Ok(())
-    }
-
+    /// Register marked complete (all required data stored locally) for this block.
     pub(super) fn register_marked_complete(
         &mut self,
         need_execution_state: bool,
-    ) -> Result<(), Error> {
+    ) -> Result<(), BlockAcquisitionError> {
         // NOTE: this is the only function wherein we should move to HaveStrictFinalitySignatures
         let new_state = match self {
             BlockAcquisitionState::HaveBlock(block, acquired_signatures, deploy_acquisition)
@@ -666,7 +810,7 @@ impl BlockAcquisitionState {
                     block.id()
                 );
                 if deploy_acquisition.needs_deploy().is_some() {
-                    return Err(Error::InvalidAttemptToMarkComplete);
+                    return Err(BlockAcquisitionError::InvalidAttemptToMarkComplete);
                 }
                 // this must be a block w/ no deploys and thus also no execution effects or
                 // approvals hashes; we can go straight to strict finality
@@ -716,165 +860,6 @@ impl BlockAcquisitionState {
         Ok(())
     }
 
-    pub(super) fn next_action(
-        &mut self,
-        peer_list: &PeerList,
-        validator_weights: &EraValidatorWeights,
-        rng: &mut NodeRng,
-        should_fetch_execution_state: bool,
-    ) -> Result<BlockAcquisitionAction, Error> {
-        let ret = match self {
-            BlockAcquisitionState::Initialized(block_hash, ..) => Ok(
-                BlockAcquisitionAction::block_header(peer_list, rng, *block_hash),
-            ),
-            BlockAcquisitionState::HaveBlockHeader(block_header, signatures) => {
-                if validator_weights.is_empty() {
-                    Ok(BlockAcquisitionAction::era_validators(
-                        validator_weights.era_id(),
-                    ))
-                } else {
-                    Ok(BlockAcquisitionAction::finality_signatures(
-                        peer_list,
-                        rng,
-                        block_header,
-                        validator_weights
-                            .missing_validators(signatures.have_signatures())
-                            .cloned()
-                            .collect(),
-                    ))
-                }
-            }
-            BlockAcquisitionState::HaveWeakFinalitySignatures(header, _) => Ok(
-                BlockAcquisitionAction::block_body(peer_list, rng, header.id()),
-            ),
-            BlockAcquisitionState::HaveBlock(block, signatures, deploy_state) => {
-                if should_fetch_execution_state {
-                    Ok(BlockAcquisitionAction::global_state(
-                        peer_list,
-                        rng,
-                        *block.hash(),
-                        *block.state_root_hash(),
-                    ))
-                } else if deploy_state.needs_deploy().is_none() {
-                    Ok(BlockAcquisitionAction::strict_finality_signatures(
-                        peer_list,
-                        rng,
-                        block.header(),
-                        validator_weights,
-                        signatures,
-                    ))
-                } else {
-                    Ok(BlockAcquisitionAction::approvals_hashes(
-                        block, peer_list, rng,
-                    ))
-                }
-            }
-            BlockAcquisitionState::HaveGlobalState(
-                block,
-                signatures,
-                deploy_state,
-                exec_results,
-            ) => {
-                if should_fetch_execution_state == false {
-                    Err(Error::InvalidStateTransition)
-                } else if deploy_state.needs_deploy().is_none() {
-                    Ok(BlockAcquisitionAction::strict_finality_signatures(
-                        peer_list,
-                        rng,
-                        block.header(),
-                        validator_weights,
-                        signatures,
-                    ))
-                } else {
-                    match exec_results {
-                        ExecutionResultsAcquisition::Needed { .. } => {
-                            Ok(BlockAcquisitionAction::execution_results_checksum(
-                                *block.hash(),
-                                *block.state_root_hash(),
-                            ))
-                        }
-                        acq @ ExecutionResultsAcquisition::Pending { .. }
-                        | acq @ ExecutionResultsAcquisition::Acquiring { .. } => {
-                            match acq.needs_value_or_chunk() {
-                                None => {
-                                    warn!(block_hash=%block.id(), "execution_results_acquisition.needs_value_or_chunk() should never be None for these variants");
-                                    Err(Error::InvalidAttemptToAcquireExecutionResults)
-                                }
-                                Some((next, checksum)) => {
-                                    Ok(BlockAcquisitionAction::execution_results(
-                                        block.id(),
-                                        peer_list,
-                                        rng,
-                                        next,
-                                        checksum,
-                                    ))
-                                }
-                            }
-                        }
-                        ExecutionResultsAcquisition::Complete { .. } => Ok(
-                            BlockAcquisitionAction::approvals_hashes(block, peer_list, rng),
-                        ),
-                    }
-                }
-            }
-            BlockAcquisitionState::HaveAllExecutionResults(
-                block,
-                signatures,
-                deploys,
-                checksum,
-            ) => match (should_fetch_execution_state, checksum) {
-                (false, _) => Err(Error::InvalidStateTransition),
-                (true, ExecutionResultsChecksum::Checkable(_)) => Ok(
-                    BlockAcquisitionAction::approvals_hashes(block, peer_list, rng),
-                ),
-                (true, ExecutionResultsChecksum::Uncheckable) => {
-                    Ok(BlockAcquisitionAction::maybe_needs_deploy(
-                        block.header(),
-                        peer_list,
-                        rng,
-                        validator_weights,
-                        signatures,
-                        deploys.needs_deploy(),
-                    ))
-                }
-            },
-            BlockAcquisitionState::HaveApprovalsHashes(block, signatures, deploys) => {
-                Ok(BlockAcquisitionAction::maybe_needs_deploy(
-                    block.header(),
-                    peer_list,
-                    rng,
-                    validator_weights,
-                    signatures,
-                    deploys.needs_deploy(),
-                ))
-            }
-            BlockAcquisitionState::HaveAllDeploys(header, signatures) => {
-                Ok(BlockAcquisitionAction::strict_finality_signatures(
-                    peer_list,
-                    rng,
-                    header.as_ref(),
-                    validator_weights,
-                    signatures,
-                ))
-            }
-            BlockAcquisitionState::HaveStrictFinalitySignatures(header, ..) => {
-                Ok(BlockAcquisitionAction::need_nothing(header.block_hash()))
-            }
-            BlockAcquisitionState::Failed(block_hash, ..) => {
-                Ok(BlockAcquisitionAction::need_nothing(*block_hash))
-            }
-        };
-        match &ret {
-            Ok(next_action) => {
-                info!(%next_action, %self);
-            }
-            Err(err) => {
-                error!(%err, %self);
-            }
-        }
-        ret
-    }
-
     fn log_finality_signature_acceptance(
         &self,
         maybe_block_hash: &Option<BlockHash>,
@@ -912,237 +897,5 @@ impl BlockAcquisitionState {
             new_state, self
         );
         *self = new_state;
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct BlockAcquisitionAction {
-    peers_to_ask: Vec<NodeId>,
-    need_next: NeedNext,
-}
-
-impl Display for BlockAcquisitionAction {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "need_next: {} from {} peers",
-            self.need_next,
-            self.peers_to_ask.len()
-        )
-    }
-}
-
-impl BlockAcquisitionAction {
-    pub(super) fn need_next(&self) -> NeedNext {
-        self.need_next.clone()
-    }
-
-    pub(super) fn peers_to_ask(&self) -> Vec<NodeId> {
-        self.peers_to_ask.to_vec()
-    }
-
-    pub(super) fn need_nothing(block_hash: BlockHash) -> Self {
-        BlockAcquisitionAction {
-            peers_to_ask: vec![],
-            need_next: NeedNext::Nothing(block_hash),
-        }
-    }
-
-    pub(super) fn peers(block_hash: BlockHash) -> Self {
-        BlockAcquisitionAction {
-            peers_to_ask: vec![],
-            need_next: NeedNext::Peers(block_hash),
-        }
-    }
-
-    pub(super) fn execution_results_checksum(
-        block_hash: BlockHash,
-        global_state_root_hash: Digest,
-    ) -> Self {
-        BlockAcquisitionAction {
-            peers_to_ask: vec![],
-            need_next: NeedNext::ExecutionResultsChecksum(block_hash, global_state_root_hash),
-        }
-    }
-
-    pub(super) fn execution_results(
-        block_hash: BlockHash,
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-        next: BlockExecutionResultsOrChunkId,
-        checksum: ExecutionResultsChecksum,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::ExecutionResults(block_hash, next, checksum),
-        }
-    }
-
-    pub(super) fn approvals_hashes(block: &Block, peer_list: &PeerList, rng: &mut NodeRng) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::ApprovalsHashes(*block.hash(), Box::new(block.clone())),
-        }
-    }
-
-    pub(super) fn deploy_by_hash(
-        block_hash: BlockHash,
-        deploy_hash: DeployHash,
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::DeployByHash(block_hash, deploy_hash),
-        }
-    }
-
-    pub(super) fn deploy_by_id(
-        block_hash: BlockHash,
-        deploy_id: DeployId,
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::DeployById(block_hash, deploy_id),
-        }
-    }
-
-    pub(super) fn global_state(
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-        block_hash: BlockHash,
-        root_hash: Digest,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::GlobalState(block_hash, root_hash),
-        }
-    }
-
-    pub(super) fn finality_signatures(
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-        block_header: &BlockHeader,
-        missing_signatures: Vec<PublicKey>,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        let era_id = block_header.era_id();
-        let block_hash = block_header.block_hash();
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::FinalitySignatures(block_hash, era_id, missing_signatures),
-        }
-    }
-
-    pub(super) fn strict_finality_signatures(
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-        block_header: &BlockHeader,
-        validator_weights: &EraValidatorWeights,
-        signature_acquisition: &SignatureAcquisition,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        let era_id = block_header.era_id();
-        let block_hash = block_header.block_hash();
-        let block_height = block_header.height();
-
-        if let SignatureWeight::Sufficient =
-            validator_weights.has_sufficient_weight(signature_acquisition.have_signatures())
-        {
-            // we have enough signatures; need to make sure we've stored the necessary bits
-            return BlockAcquisitionAction {
-                peers_to_ask: vec![],
-                need_next: NeedNext::BlockMarkedComplete(block_hash, block_height),
-            };
-        }
-        // need more signatures
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::FinalitySignatures(
-                block_hash,
-                era_id,
-                validator_weights
-                    .missing_validators(signature_acquisition.have_signatures())
-                    .cloned()
-                    .collect(),
-            ),
-        }
-    }
-
-    pub(super) fn block_body(
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-        block_hash: BlockHash,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::BlockBody(block_hash),
-        }
-    }
-
-    pub(super) fn block_header(
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-        block_hash: BlockHash,
-    ) -> Self {
-        let peers_to_ask = peer_list.qualified_peers(rng);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next: NeedNext::BlockHeader(block_hash),
-        }
-    }
-
-    pub(super) fn era_validators(era_id: EraId) -> Self {
-        let peers_to_ask = Default::default();
-        let need_next = NeedNext::EraValidators(era_id);
-        BlockAcquisitionAction {
-            peers_to_ask,
-            need_next,
-        }
-    }
-
-    pub(super) fn maybe_needs_deploy(
-        block_header: &BlockHeader,
-        peer_list: &PeerList,
-        rng: &mut NodeRng,
-        validator_weights: &EraValidatorWeights,
-        signatures: &SignatureAcquisition,
-        needs_deploy: Option<DeployIdentifier>,
-    ) -> Self {
-        match needs_deploy {
-            Some(DeployIdentifier::ById(deploy_id)) => {
-                info!("BlockAcquisition: requesting missing deploys by ID");
-                BlockAcquisitionAction::deploy_by_id(
-                    block_header.block_hash(),
-                    deploy_id,
-                    peer_list,
-                    rng,
-                )
-            }
-            Some(DeployIdentifier::ByHash(deploy_hash)) => {
-                info!("BlockAcquisition: requesting missing deploys by hash");
-                BlockAcquisitionAction::deploy_by_hash(
-                    block_header.block_hash(),
-                    deploy_hash,
-                    peer_list,
-                    rng,
-                )
-            }
-            None => BlockAcquisitionAction::strict_finality_signatures(
-                peer_list,
-                rng,
-                block_header,
-                validator_weights,
-                signatures,
-            ),
-        }
     }
 }
